data_pipeline/requirements.txt
# Data Downloaders
alpaca-trade-api
python-binance
pandas
numpy
requests
yfinance
python-dateutil
pytz
tqdm
polygon-api-client
databento
investpy
pandas_datareader
alpha-vantage
quandl
tiingo
fredapi
beautifulsoup4
lxml
html5lib
nsepy
selenium
webdriver-manager
semantic-kernel
openai
scipy
scikit-learn
matplotlib
seaborn
plotly
pyarrow
fastparquet
tables
h5py
asyncio-mqtt
aiofiles
uvloop
python-decouple
pydantic
pydantic-settings
rich
typer
click
colorama